llvm interp: cleanup and test case removal
The test case that fails here (gep7) also fails in the reference
implementation. The test access a struct array member out of bounds,
this would mean that the pointer moves to the next field in the outer
struct, but that logic is not baked in to the reference interpreter, so
that test fails when we interpret it. Maybe one day we will add support
for out of bounds accesses.