custom parsing return.ll!
the bare minimum! but the code is looking alright.
going for a functional style so i made sure the ctx was copy so we can
backtrack easily. we'll see if this is too much of a pain and if we even
need the lookahead in the case of parsing ll code, but for now it's
looking well.