x64 ast uses usize labels, not intern strs
we keep a top level vec mapping ids to label names and we just use
indicies. we have to build a hashmap to map from name to index while
creating them, but we discard it afterward. this removes the x64
backend's dependency on internment, and we should probably do the same
for the other stages of the compiler cause it's simpler and smaller.